Moved plotting functions from predict.py to new utils.py
